requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
httpx
python-dotenv
asyncpg
sqlalchemy
supabase
python-dotenv